fgets returning value and char-subscript fixed

+++ lab1-0/src/BoyerMoore.c
@@ -38,7 +38,7 @@ void strToSearchState(const char* str, BMSearchState* state) {
 	}
 	i = 0;
 	while (i < len - 1) {
-		shift[str[i]] = len - i - 1;
+		shift[(unsigned int)str[i]] = len - i - 1;
 		++i;
 	}
 }
@@ -66,8 +66,8 @@ unsigned int findSubString(const char* sample, const char* text, unsigned int te
 			globalIdx += state.len;
 		}
 		else {
-			globalIdx += state.shift[text[sampleIdx]];
-			sampleIdx += state.shift[text[sampleIdx]];
+			globalIdx += state.shift[(unsigned int)text[sampleIdx]];
+			sampleIdx += state.shift[(unsigned int)text[sampleIdx]];
 		}
 	}
 	return globalIdx;
